requirements.txt
pillow
scipy
deskew
pytesseract
pysimplegui
pygame
textblob
spacy
contextualSpellCheck
gtts
tts
paddlepaddle
shapely
scikit-image
imgaug
pyclipper
lmdb
tqdm
numpy
visualdl
rapidfuzz
opencv-python==4.6.0.66
opencv-contrib-python==4.6.0.66
cython
lxml
premailer
openpyxl
attrdict
Polygon3
lanms-neo==1.0.2
PyMuPDF<1.21.0

